refactored benchmark into its own test
